Fix reading float data as integer before converting to long. Bump version to 1.1.5.

+++ MVLibraryNET/MBE/Sheet.cs
@@ -146,7 +146,7 @@ public unsafe class Sheet
                         break;
                     case ColumnType.Float:
                         Utils.Align(ref rowCellOffset, 4);
-                        cellValue = BitConverter.ToInt32(rowBuffer, rowCellOffset);
+                        cellValue = (long)BitConverter.ToSingle(rowBuffer, rowCellOffset);
                         rowCellOffset += 4;
                         break;
                     case ColumnType.String:
